requirements.txt
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
wheel==0.37.1
Pillow==9.1.1
email-validator==1.2.1
itsdangerous==2.0.1